Add files via upload
so i mocked this up as a rough copy of the text-based version of our game for monday's class. It compiles but the logic doesn't quite work. I don't think i'll have any time to work on it between now and class. sorry. hope this works well-enough for a skeleton.